requirements.txt
torch>=1.5.1
pandas>=1.0.5
numpy>=1.18.5
nbformat>=5.0.7
pyyaml>=5.3.1
sklearn
argparse>=1.4.0
matplotlib>=3.2.2
loompy>=3.0.6
tqdm>=4.46.1
anndata>=0.7.3
rootpath>=0.1.1
nbconvert>=5.6.1
nbformat>=5.0.7
seaborn>=0.10.1
jupyter_client>=6.1.3
ipykernel>=5.3.0
ipython>=7.15.0
h5py>=2.10.0
tbb>=2019.5
FlowCytometryTools>=0.5.0
fastcluster>=1.1.26
sphinx>=3.1.2
sphinx_autodoc_typehints>=1.11.0
sphinx_rtd_theme>=0.5.0
autodocsumm>=0.1.13
nbsphinx>=0.7.1
nbsphinx_link>=1.3.0
tables>=3.6.1
typeguard
mypy
